[Refactor] Move HTTP auth helpers into HttpCore
