+++ be/src/http/CMakeLists.txt
@@ -19,6 +19,7 @@ set(EXECUTABLE_OUTPUT_PATH "${BUILD_DIR}/src/http")
 
 set(HTTP_CORE_FILES
   http_metrics.cpp
+  http_auth.cpp
   http_headers.cpp
   http_method.cpp
   http_request.cpp
