+++ be/src/http/action/stream_load.cpp
@@ -60,12 +60,12 @@
 #include "gen_cpp/FrontendService.h"
 #include "gen_cpp/FrontendService_types.h"
 #include "gen_cpp/HeartbeatService_types.h"
+#include "http/http_auth.h"
 #include "http/http_channel.h"
 #include "http/http_common.h"
 #include "http/http_headers.h"
 #include "http/http_request.h"
 #include "http/http_response.h"
-#include "http/utils.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/batch_write/batch_write_util.h"
 #include "runtime/client_cache.h"
