+++ be/test/CMakeLists.txt
@@ -254,7 +254,6 @@ set(EXEC_FILES
         ./http/ev_http_server_test.cpp
         ./http/http_metrics_test.cpp
         ./http/http_client_test.cpp
-        ./http/http_utils_test.cpp
         ./http/message_body_sink_test.cpp
         ./http/metrics_action_test.cpp
         ./http/stream_load_test.cpp
