+++ be/test/http/http_client_test.cpp
@@ -24,10 +24,10 @@
 #include "common/config_storage_fwd.h"
 #include "common/logging.h"
 #include "http/ev_http_server.h"
+#include "http/http_auth.h"
 #include "http/http_channel.h"
 #include "http/http_handler.h"
 #include "http/http_request.h"
-#include "http/utils.h"
 
 namespace starrocks {
